# HelpDeskPro/postman/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "HelpDeskPro API",
    "description": "API para gestión de tickets de soporte",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7198",
      "description": "Dev"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica a un usuario y devuelve el token JWT.",
        "requestBody": {
          "description": "Credenciales de inicio de sesión",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticación correcta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresca el token JWT usando el refresh token.",
        "requestBody": {
          "description": "Solicitud con refresh token",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refrescado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Refresh token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra un nuevo usuario y devuelve sus tokens de acceso.",
        "requestBody": {
          "description": "Datos del usuario a registrar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario registrado y autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de registro inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/get-all": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Obtiene usuarios paginados y filtrados.",
        "requestBody": {
          "description": "Filtros y paginación",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUsersFiltersInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Listado paginado de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUserDtoGenericPaginationOutputDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtiene un usuario por su identificador.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador del usuario",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailUserDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/by-email/{email}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Obtiene un usuario por su email.",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "Email del usuario",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailUserDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/create": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Crea un usuario.",
        "requestBody": {
          "description": "Datos del nuevo usuario",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailUserDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos de entrada inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get-grouped-by-role": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Obtiene usuarios agrupados por rol.",
        "requestBody": {
          "description": "Paginación",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Listado de roles con usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListRolesWithUsersDtoGenericPaginationOutputDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/get-grouped-by-team": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Obtiene usuarios agrupados por equipo.",
        "requestBody": {
          "description": "Paginación",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginationInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Listado de equipos con usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsersGroupedByTeamDtoGenericPaginationOutputDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddUserDto": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario.",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario.",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario.",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos del Usuario.",
            "nullable": true
          },
          "languageCode": {
            "type": "string",
            "description": "Código de idioma preferido (opcional).",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/Roles"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si el usuario se crea activo."
          }
        },
        "additionalProperties": false,
        "description": "Datos para crear usuario por un administrador."
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Token de acceso JWT.",
            "nullable": true
          },
          "refreshToken": {
            "type": "string",
            "description": "Token de refresco.",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "description": "Fecha y hora de expiración del token de acceso.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Respuesta de autenticación con tokens de acceso y refresco."
      },
      "DetailUserDto": {
        "required": [
          "email",
          "firstName",
          "languageCode",
          "lastName",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador del usuario.",
            "format": "int32"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico.",
            "nullable": true
          },
          "lastLoginAt": {
            "type": "string",
            "description": "Fecha/hora del último inicio de sesión.",
            "format": "date-time",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Nombre.",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos.",
            "nullable": true
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario.",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si el usuario está activo."
          },
          "createdAt": {
            "type": "string",
            "description": "Fecha de creación.",
            "format": "date-time"
          },
          "languageCode": {
            "type": "string",
            "description": "Código de idioma preferido.",
            "nullable": true
          },
          "teamNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Equipos a los que pertenece.",
            "nullable": true
          },
          "tickets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListTicketDto"
            },
            "description": "Tickets creados por el usuario.",
            "nullable": true
          },
          "assignedTickets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListTicketDto"
            },
            "description": "Tickets asignados al usuario.",
            "nullable": true
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListTicketCommentDto"
            },
            "description": "Comentarios del usuario.",
            "nullable": true
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListTicketAttachmentDto"
            },
            "description": "Adjuntos subidos por el usuario.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Usuario con información detallada y relaciones."
      },
      "GetUsersFiltersInputDto": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "description": "Página solicitada (1-based).",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Tamaño de página.",
            "format": "int32"
          },
          "exactMatch": {
            "type": "boolean",
            "description": "Si es true, los filtros se aplican en coincidencia exacta."
          },
          "id": {
            "type": "integer",
            "description": "Identificador del usuario.",
            "format": "int32",
            "nullable": true
          },
          "email": {
            "type": "string",
            "description": "Email del usuario.",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario.",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos del usuario.",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Roles a incluir en el filtro.",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado de actividad del usuario.",
            "nullable": true
          },
          "createdAfterDate": {
            "type": "string",
            "description": "Fecha de creación posterior a.",
            "format": "date-time",
            "nullable": true
          },
          "createdBeforeDate": {
            "type": "string",
            "description": "Fecha de creación anterior a.",
            "format": "date-time",
            "nullable": true
          },
          "lastLoginAfterDate": {
            "type": "string",
            "description": "Último login posterior a.",
            "format": "date-time",
            "nullable": true
          },
          "lastLoginBeforeDate": {
            "type": "string",
            "description": "Último login anterior a.",
            "format": "date-time",
            "nullable": true
          },
          "teamIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": "Ids de equipos a filtrar.",
            "nullable": true
          },
          "teamNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Nombres de equipos a filtrar.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Filtros disponibles para búsqueda de usuarios con paginación."
      },
...
      "ListRolesWithUsersDtoGenericPaginationOutputDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListRolesWithUsersDto"
            },
            "description": "Elementos de la página actual.",
            "nullable": true
          },
          "paginationMetadata": {
            "$ref": "#/components/schemas/PaginationOutputDto"
          }
        },
        "additionalProperties": false,
        "description": "Resultado paginado genérico."
      },
...
      "ListUserDtoGenericPaginationOutputDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListUserDto"
            },
            "description": "Elementos de la página actual.",
            "nullable": true
          },
          "paginationMetadata": {
            "$ref": "#/components/schemas/PaginationOutputDto"
          }
        },
        "additionalProperties": false,
        "description": "Resultado paginado genérico."
      },
...
      "ListUsersGroupedByTeamDtoGenericPaginationOutputDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ListUsersGroupedByTeamDto"
            },
            "description": "Elementos de la página actual.",
            "nullable": true
          },
          "paginationMetadata": {
            "$ref": "#/components/schemas/PaginationOutputDto"
          }
        },
        "additionalProperties": false,
        "description": "Resultado paginado genérico."
      },
      "LoginRequestDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario.",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario.",
            "nullable": true
          },
          "languageCode": {
            "type": "string",
            "description": "Código de idioma preferido (opcional).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Datos de entrada para iniciar sesión."
      },
      "PaginationInputDto": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "description": "Página solicitada (1-based).",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "Tamaño de página.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Parámetros de entrada para paginación."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RefreshTokenRequestDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token emitido previamente.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Solicitud para refrescar el token de acceso."
      },
      "RegisterUserRequestDto": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email del usuario.",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario.",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario.",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos del Usuario.",
            "nullable": true
          },
          "languageCode": {
            "type": "string",
            "description": "Código de idioma preferido (opcional).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Datos de entrada para registrar un nuevo usuario."
      },
      "Roles": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Inserta el token JWT con el prefijo 'Bearer '. Ejemplo: 'Bearer eyJhbGciOi...'.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}